move declarations to own file
