+++ compressor/constants.py
@@ -1,0 +1,13 @@
+"""Constant definitions are defined here, and imported everywhere else in the
+code.
+
+This should contain definitions for values that do not change while the program
+is running, and are to remain constant throughout the execution
+"""
+VERSION = '0.1.0'
+
+ENC = 'utf-8'
+BYTE = 8
+BUFF_SIZE = 1024
+LEFT = b'0'
+RIGHT = b'1'
