+++ doc/conf.py
@@ -25,7 +25,7 @@ import sphinx_rtd_theme
 
 sys.path.insert(0, os.path.abspath('../..'))
 
-from compressor import VERSION
+from compressor.constants import VERSION
 # -- General configuration ------------------------------------------------
 
 # If your documentation needs a minimal Sphinx version, state it here.
